requirements.txt
numpy>=1.16.5
Pillow>=6.1.0
tensorflow>=1.14.0
